feat(脚本): 完善自定义证书提示

+++ install.sh
@@ -1755,11 +1755,13 @@ installTLS() {
             fi
 
         else
-            echoContent yellow " ---> 如未过期或者自定义证书请选择[n]\n"
-            read -r -p "是否重新安装？[y/n]:" reInstallStatus
-            if [[ "${reInstallStatus}" == "y" ]]; then
-                rm -rf /etc/v2ray-agent/tls/*
-                installTLS "$1"
+            if [[ -d "$HOME/.acme.sh/${tlsDomain}_ecc" && -f "$HOME/.acme.sh/${tlsDomain}_ecc/${tlsDomain}.key" && -f "$HOME/.acme.sh/${tlsDomain}_ecc/${tlsDomain}.cer" ]] || [[ "${installedDNSAPIStatus}" == "true" ]]; then
+                echoContent yellow " ---> 如未过期或者自定义证书请选择[n]\n"
+                read -r -p "是否重新安装？[y/n]:" reInstallStatus
+                if [[ "${reInstallStatus}" == "y" ]]; then
+                    rm -rf /etc/v2ray-agent/tls/*
+                    installTLS "$1"
+                fi
             fi
         fi
 
@@ -2057,6 +2059,8 @@ renewalTLS() {
         else
             echoContent green " ---> 证书有效"
         fi
+    elif [[ -f "/etc/v2ray-agent/tls/${tlsDomain}.crt" && -f "/etc/v2ray-agent/tls/${tlsDomain}.key" && -n $(cat "/etc/v2ray-agent/tls/${tlsDomain}.crt") ]]; then
+        echoContent yellow " ---> 检测到使用自定义证书，无法执行renew操作。"
     else
         echoContent red " ---> 未安装"
     fi
@@ -9139,7 +9143,7 @@ menu() {
 	echoContent red "\n=============================================================="
 	echoContent green "原作者：mack-a"
 	echoContent green "作者：Wizard89"
-	echoContent green "当前版本：v3.0.47"
+	echoContent green "当前版本：v3.0.48"
 	echoContent green "Github：https://github.com/Wizard89/v2ray-agent"
 	echoContent green "描述：八合一共存脚本\c"
     showInstallStatus
